First rudimentary sketch
Committing also files from 2002 and 2003 which serve as guidance to
create this new Clonk object.

Studied the relevant shape boundaries in Study [2021-11-24_1820].

Developed a first Graphics draft in Study [2021-11-24_1843].

Tried to put together a simple toy object to make sure the object can be
built in tunnels, this is Clonk Object [2021-11-26_1846].  For some
reason, the engine complains that the object "cannot be built" when
trying to build the object using a toolkit (N.B. that this is different
from "not enough space" or other limitations).