Understood the graphics determination a little better
During construction, it seems the graphic is determined soleley by the
``Width`` and ``Height`` DefCore values.  They are just the top-left
corner of the Graphics file.

Once the object has been constructed, the Facet can be changed by means
of an Action.  This can be started by ``SetAction``.  A Delay of 0 means
that the action never ends.  Otherwise, the Action returns to Idle once
it completes (given no NextAction value).

For me, this means that:

1.  I need to provide a 22 px high base graphic in the left-top corner
    of the Graphics.png.

2.  Once the object has been built, a new Action needs to be fired up
    which includes the top pipe and the bottom floor, which extend into
    the tunnel ceiling/floor.

By this, the ceiling and the floor will become visible after
construction of the airlock.