Got the graphics working
During construction, the pipe in the ceiling and the floor are clipped
off, as they penetrate into the tunnel ceiling/floor.  Once constructed,
an Action "Trivial" kicks in, showing the complete graphics.

Missing pieces:

*   Proper Vertices;
*   Proper offset of the "extended" graphics.

During this, one needs to bear in mind that the "official" height +
width are smaller than the "extended" graphics; this will have
implications for the general offset (during construction) and for the
coordinates of the vertices.