On the way to providing proper vertices
With the help of a detailed draft on paper I successfully provided a
Clonk Object with a graphics reaching into the ceiling and the floor.

When designing the coordinates of the vertices, I assumed that they
would touch the ground by being just one pixel above the ground; this
turned out as wrong.  Instead, they are sticking in the first row of the
ground region.

The next step is thus to update the vertex coordinates to this fact.