Setting up the Object to have the actual width of the final Graphics
The ``Width`` of the Object has been increased from 169 px to 173 px, to
include also the external features of the final Graphics (with pipes and
floor).  The ``Offset`` needed to be increased in absolute value by 2
px, and in the ``Facet`` of the 'Tivial' Action the horizontal offset
needed to be set to ``0`` (it was ``-2`` before).