Providing an Idle Frame at the beginning of some animations
Since animations should be *reversible* (by setting ``Reverse=1`` in the
ActMap.txt), not only the *last frame* should be the Idle state, but
also the *first* frame.

This applies to the Shield Opening Animations as well as to the present
Walk Animation.  The Pump Left / Pump Right Animations are cyclic and
need no change here.