First attempts towards a control menu
I used :func:`ControlDown` to close both shields, however, this is not a
good idea since it renders releasing the grabbed Airlock impossible.

It would be feasible to group *both* opening *and* closing in *one*
menu, opened by :func:`ControlDown`.  Possibly it would be a good idea
to change the scope of the menus to "anything regarded the
left/right/both door(s)".