Eliminated *Done EndCall fields of the Shield* Actions
Instead of deciding which Action to take next to reach the Aim at the
end of the Shield* Actions by means of EndCall *Done Callbacks, this
decision is now left to the Run function, and the Shield* Actions lead
to the appropriate infinite Action.  This has two advantages:

1.  There are fewer infinite Actions reached by the Shield* Actions than
    there are Shield* Actions, so the Run function turns out shorter
    than all the *Done EndCall functions.

2.  The Idling* Action can install the right SolidMap via their
    StartCall.

This approach to installing the SolidMap implies that most probably
there will be *no* intermediate SolidMaps for opening/closing the
Shields, and the SolidMap will change always at the *end* of the Shield*
Actions by the StartCall of the next Idling* Action.  Providing an
"animated" SolidMap would require corresponding SolidMap animations for
*each and every* Shield* Action.