Work on the "manual" Aims
Removed the Aim ``doTransfer``.  It has been replaced by the
``transfer`` Mode.  There will be a multitude of Aims in the
``transfer`` Mode.

Added structure to :func:`RunOpenedBoth` to react on the cautious Aims.
Here more elaborate code will be required to "be cautious".

Added :func:`ControlDig` which selects the mode of operation.  The lines
initiating the transition to Transfer Mode and Drypumping Mode are
currently commented out since these modes are entirely unimplemented at
the moment.

Mode transitions between Cautious and Forced mode are present.