The Airlock turns alpha
The *top* variable needed adjustment from ``-10`` to ``-9``, otherwise
the topmost filling degree never was recognised.  it should be ``-10``,
I cannot spot the mistake.  It does not matter much, though.

There is some insignificant problem with pumping: After the pump
extraction point turned dry, and the pump is turned off, there appears
liquid again the next frame (or so).  There is monitoring code in the
Run function which reveals this.  This results in unnecessary pumping
the next time the centre is required dry.  The present code contains
some commented-out approaches to solve this problem.

Furthermore a small behavioural drawback appears when changing the *aim*
from ``transferLeft`` to ``transferRight`` (or vice versa) while the
pump is acting.  In this case, the pumping does not stop even when it
would be permissible without problems.  It is possible to first put the
Airlock into the waitForTransfer aim and then instructing the new
transfer aim again.  It is not yet clear how to improve on this.