Iterating all floor pixels for pumping
This commit fixes the issue reported in 9794a7ca regarding incomplete
result of the pumping operation.  It achieves this by pumping out not
only at the centre of the airlock, but also on all other floor pixels
between the shields.

For decisions about whether the floor is *dry*, *only* the centre pixel
is checked.

Most tracking code related to this issue has been commented out.