Reviewed the new Script
The following procedure has proven useful w.r.t. coding *state changes*:

1.  When the state change can be handled by setting a new aim, do that.
2.  Otherwise, if a change of Action is necessary to *reach* the
    *present aim*, instruct this Action change.

In this way, the amout of Action changes is reduced to a minimum, while
maintaining (hopefully) fully automatic behaviour.

I observed that often one is "tempted" to code an Action change which is
anticipating future necessities.  However, I think that it is better to
leave that decision to the code dealing with the next aim.

One should *never* need to code *both* an aim change *together* with an
Action change at the same time.

With this commit the ``doTransferLeft`` and ``doTransferRight`` aims
have been replaced by a more granular set of aims: ``{enterFrom,
exitTo}{Left, Right}``.  This has been inspired by the fact that it
makes it possible to react to change of operation when a passenger
changes *direction* while in the central region of the airlock.  Before,
it's been impossible to detect that the passenger actually wanted to
*leave* the airlock from the *centre*, instead of attempting to *enter*
the airlock from the other side.  Before, the pumping was initiated
because the passenger wanted to *leave* the airlock to that side.

I feel that the new aim options *might* increase the "intelligence" of
the airlock's program, but I want to see it working before.

Another option might have been to make use of the Count* functions in
the Run function of the Pump* Actions.  The present approach however
appears to eliminate the ambiguity described above more at its root
position.

I've read through the whole code carefully and have hope that it might
work.

The more fine-grained set of aims also made it possible to detect
passenger-less situations; in this case the aim is set to
``waitForTransfer`` now.