Implemented Dry Pumping
There was need for another Aim: ``idleDryPumping``.  It is required
because in dry pumping mode the *direction* isn't always defined.

Bugfixes in :func:`PumpOut`: I forgot to take the *xcentre* into account
in several places, leading to various dysfunctions of the airlock in dry
pump mode.

The dry pump mode now is accessible via the Dig Control.

I improved the transition behaviour from transfer mode to
cautious/forced mode.  The translation now is more detailed.  This is
another benefit of the more fine-grained transfer mode aims.