The current Script is to be released
I've read through the whole script step by step and verified that each
decision is sound and that situation symmetric to each other are handled
analoguous.

The only substantial change was in RunClosedLeft/exitToLeft resp.
RunClosedRight/exitToRight: Here, when the passengers are in the resp.
outer region, the airlock now does plain *nothing* (before, it closed
the shield at the opposite site).

Noteworthy is also that the RunPump* Actions now are cancelled in
drypumping mode.

Some cosmetic changes w.r.t. ordering of sections for the different
modes.  Also a few updates of comments.